# src/main/api/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Flowrider API
  description: |
    REST API for Flowrider - AI Session Orchestration Platform

    Flowrider is an Electron app for orchestrating multiple AI terminal sessions using a 3D icosahedron interface (20 faces = 20 sessions).

    ## Authentication
    All API requests require authentication using an API key passed in the `X-API-Key` header.

    ## Rate Limiting
    API requests are rate-limited to 100 requests per minute per API key.

    ## Error Handling
    The API uses standard HTTP status codes and returns error details in JSON format.
  version: 1.0.0
  contact:
    name: Josephas LLC
    url: https://github.com/josephas-llc/flowrider
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3737/api
    description: Local development server
  - url: https://api.flowrider.local/api
    description: Production server (when deployed)
...
security:
  - ApiKeyAuth: []
paths:
...
  /projects:
    get:
      tags:
        - Projects
      summary: List all projects
      description: Retrieve a list of all projects
      operationId: listProjects
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Project'
                  total:
                    type: integer
                    example: 15
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags:
        - Projects
      summary: Create a new project
      description: Create a new project workspace
      operationId: createProject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - path
              properties:
                name:
                  type: string
                  example: "E-commerce Platform"
                path:
                  type: string
                  description: Local file system path
                  example: "/Users/dev/ecommerce"
                description:
                  type: string
                  example: "Full-stack e-commerce platform with Next.js and Stripe"
                repoUrl:
                  type: string
                  format: uri
                  example: "https://github.com/myorg/ecommerce"
                language:
                  type: string
                  example: "typescript"
                tags:
                  type: array
                  items:
                    type: string
                  example: ["nextjs", "stripe", "postgresql"]
      responses:
        '201':
          description: Project created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Project'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /projects/{id}:
    get:
      tags:
        - Projects
      summary: Get project details
      description: Retrieve details for a specific project
      operationId: getProject
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Project'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      tags:
        - Projects
      summary: Update project
      description: Update project details
      operationId: updateProject
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: "Updated Project Name"
                description:
                  type: string
                path:
                  type: string
                repoUrl:
                  type: string
                  format: uri
                language:
                  type: string
                tags:
                  type: array
                  items:
                    type: string
      responses:
        '200':
          description: Project updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Project'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      tags:
        - Projects
      summary: Delete project
      description: Delete a project
      operationId: deleteProject
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      responses:
        '200':
          description: Project deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Project deleted successfully"
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # ========================================
  # WEBHOOKS
  # ========================================
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication
  parameters:
...
    ProjectId:
      name: id
      in: path
      required: true
      description: Project ID
      schema:
        type: string
        example: "proj_123abc"
...
  schemas:
...
    Project:
      type: object
      properties:
        id:
          type: string
          example: "proj_123abc"
        name:
          type: string
          example: "E-commerce Platform"
        path:
          type: string
          example: "/Users/dev/ecommerce"
        description:
          type: string
          nullable: true
          example: "Full-stack e-commerce platform"
        repoUrl:
          type: string
          format: uri
          nullable: true
          example: "https://github.com/myorg/ecommerce"
        language:
          type: string
          nullable: true
          example: "typescript"
        tags:
          type: array
          items:
            type: string
          example: ["nextjs", "stripe", "postgresql"]
        sessionCount:
          type: integer
          description: Number of active sessions for this project
          example: 3
        createdAt:
          type: string
          format: date-time
          example: "2026-07-01T10:00:00Z"
        updatedAt:
          type: string
          format: date-time
          example: "2026-07-06T15:30:00Z"
...
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: "Invalid request parameters"
        code:
          type: string
          description: Error code
          example: "INVALID_INPUT"
        details:
          type: object
          nullable: true
          description: Additional error details
  responses:
    BadRequestError:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error: "Invalid request parameters"
            code: "INVALID_INPUT"
    UnauthorizedError:
      description: Unauthorized - Invalid or missing API key
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error: "Invalid or missing API key"
            code: "UNAUTHORIZED"
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error: "Resource not found"
            code: "NOT_FOUND"
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error: "An internal server error occurred"
            code: "INTERNAL_ERROR"